fix unit test issues
