+++ rag_experiment_accelerator/checkpoint/tests/test_null_checkpoint.py
@@ -1,6 +1,8 @@
 import unittest
 
+from rag_experiment_accelerator.checkpoint.checkpoint_factory import CheckpointFactory
 from rag_experiment_accelerator.checkpoint.null_checkpoint import NullCheckpoint
+from rag_experiment_accelerator.config.config import ExecutionEnvironment
 
 
 def dummy(word):
@@ -9,7 +11,10 @@ def dummy(word):
 
 class TestNullCheckpoint(unittest.TestCase):
     def test_wrapped_method_is_not_cached(self):
-        checkpoint = NullCheckpoint()
+        checkpoint = CheckpointFactory.create_checkpoint(
+            ExecutionEnvironment.LOCAL, False
+        )
+        self.assertIsInstance(checkpoint, NullCheckpoint)
         data_id = "unique_id"
         result1 = checkpoint.load_or_run(dummy, data_id, "first run")
         result2 = checkpoint.load_or_run(dummy, data_id, "second run")
