+++ rag_experiment_accelerator/run/tests/test_querying.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock, patch
 from azure.search.documents import SearchClient
 from rag_experiment_accelerator.checkpoint import CheckpointFactory
 from rag_experiment_accelerator.embedding.embedding_model import EmbeddingModel
-from rag_experiment_accelerator.config.config import Config
+from rag_experiment_accelerator.config.config import Config, ExecutionEnvironment
 from rag_experiment_accelerator.config.index_config import IndexConfig
 from rag_experiment_accelerator.config.environment import Environment
 from rag_experiment_accelerator.run.querying import (
@@ -310,6 +310,7 @@ class TestQuerying(unittest.TestCase):
         )
         mock_config.index_configs.return_value = [index_config]
         mock_config.use_checkpoints = False
+        mock_config.execution_environment = ExecutionEnvironment.LOCAL
         CheckpointFactory.create_checkpoint(
             mock_config.execution_environment, mock_config.use_checkpoints, ""
         )
