+++ rag_experiment_accelerator/sampling/tests/test_clustering.py
@@ -3,7 +3,9 @@ import pandas as pd
 import pytest
 from unittest.mock import MagicMock, patch
 
+from rag_experiment_accelerator.config.config import ExecutionEnvironment
 from rag_experiment_accelerator.sampling.clustering import cluster, load_parser
+from rag_experiment_accelerator.checkpoint import CheckpointFactory
 
 
 @pytest.fixture
@@ -93,6 +95,8 @@ def test_cluster(mock_logger, mock_df, mock_reducer, mock_df_concat, mock_data_d
         config.sampling_output_dir,
         f"sampled_cluster_predictions_cluster_number_{config.sample_optimum_k}.csv",
     )
+    config.execution_environment = ExecutionEnvironment.LOCAL
+    CheckpointFactory.create_checkpoint(config.execution_environment, False)
     os.makedirs(config.sampling_output_dir)
 
     with patch(
